requirements.txt
appdirs==1.4.0
avro==1.8.1
docopt==0.6.2
kafka-python==1.3.2
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
PyYAML==3.12
six==1.10.0
